carbon changes. r=pink/sdagley/beard. a=sfraser.

+++ widget/src/mac/nsMacTSMMessagePump.h
@@ -45,9 +45,9 @@ public:
 	void Shutdown();
 	
 private:
-	static pascal OSErr PositionToOffsetHandler(const AppleEvent *theAppleEvent, AppleEvent *reply, UInt32 handlerRefcon);
-	static pascal OSErr OffsetToPositionHandler(const AppleEvent *theAppleEvent, AppleEvent *reply, UInt32 handlerRefcon);
-	static pascal OSErr UpdateHandler(const AppleEvent *theAppleEvent, AppleEvent *reply, UInt32 handlerRefcon);
+	static pascal OSErr PositionToOffsetHandler(const AppleEvent *theAppleEvent, AppleEvent *reply, long handlerRefcon);
+	static pascal OSErr OffsetToPositionHandler(const AppleEvent *theAppleEvent, AppleEvent *reply, long handlerRefcon);
+	static pascal OSErr UpdateHandler(const AppleEvent *theAppleEvent, AppleEvent *reply, long handlerRefcon);
 	static AEEventHandlerUPP mPos2OffsetUPP;
 	static AEEventHandlerUPP mOffset2PosUPP;
 	static AEEventHandlerUPP mUpdateUPP;
